3.IActionResult ve Action Overloading

+++ HotelFinder.API/Controllers/HotelsController.cs
@@ -23,29 +23,52 @@ public class HotelsController : ControllerBase
     /// </summary>
     /// <returns></returns>
     [HttpGet]
-    public List<Hotel> Get(){
-        return this.hotelService.GetAllHotels();
+    public IActionResult Get(){
+        var hotels = this.hotelService.GetAllHotels();
+        return Ok(hotels); // 200 + body kısmına hotels'i ekle
     }
 
     [HttpGet]
-    [Route("{id}")]
-    public Hotel Get(int id){
-        return this.hotelService.GetHotelById(id);
+    [Route("GetHotelById/{id}")]
+    public IActionResult Get(int id){
+        var hotel = this.hotelService.GetHotelById(id);
+        if (hotel != null){
+            return Ok(hotel); // 200 + data
+        }
+        return NotFound();
+    }
+
+    [HttpGet]
+    [Route("[action]/{name}")]
+    public IActionResult GetHotelByName(string name){
+        var hotel = this.hotelService.GetHotelByName(name);
+        if (hotel != null){
+            return Ok(hotel);
+        }
+        return NotFound();
     }
 
     [HttpPost]
-    public Hotel Post([FromBody] Hotel hotel){
-        return this.hotelService.CreateHotel(hotel);
+    public IActionResult Post([FromBody] Hotel hotel){
+        var createdHotel = this.hotelService.CreateHotel(hotel);
+        return CreatedAtAction("Get",new {id=createdHotel.Id},createdHotel); // 201 + data
     }
     
     [HttpPut]
-    public Hotel Put([FromBody] Hotel hotel){
-        return this.hotelService.UpdateHotel(hotel);
+    public IActionResult Put([FromBody] Hotel hotel){
+        if (this.hotelService.GetHotelById(hotel.Id) != null){
+            return Ok(this.hotelService.UpdateHotel(hotel));
+        }
+        return NotFound();
     }
 
     [HttpDelete]
     [Route("{id}")]
-    public void Delete(int id){
-        this.hotelService.DeleteHotel(id);
+    public IActionResult Delete(int id){
+        if (this.hotelService.GetHotelById(id) != null){
+            this.hotelService.DeleteHotel(id);
+            return Ok();
+        }
+        return NotFound();
     }
 }
